Adjust abstract to interface and update package
Changed abstract classes RadioCommandExecutorFactory and
RadioUnitReceiverFactory to interfaces. Moved classes to
RadioUnit package.
