+++ src/main/java/radiounit/EricssonWcdmaRadioUnitReceiver.java
@@ -1,4 +1,4 @@
-package radiounit_executorandreceiver;
+package radiounit;
 
 import java.util.List;
