+++ src/main/java/radiounit/RadioCommandExecutorFactory.java
@@ -1,0 +1,12 @@
+package radiounit;
+
+/**
+ * Abstract class for RadioCommandExecutorFactory classes
+ * @author esiumat
+ *
+ */
+public interface RadioCommandExecutorFactory {
+
+	public RadioCommandExecutor createRadioCommandExecutor();
+	
+}
