+++ src/main/java/radiounit/RadioUnitReceiverFactory.java
@@ -1,0 +1,12 @@
+package radiounit;
+
+/**
+ * Abstract class for RadioUnitReceiverFactory classes
+ * 
+ * @author esiumat
+ *
+ */
+public interface RadioUnitReceiverFactory {
+
+	public RadioUnitReceiver createRadioUnitRecevier();
+}
