--- src/main/java/radiounit_executorandreceiver/RadioCommandExecutorFactory.java
@@ -1,12 +1,0 @@
-package radiounit_executorandreceiver;
-
-/**
- * Abstract class for RadioCommandExecutorFactory classes
- * @author esiumat
- *
- */
-public abstract class RadioCommandExecutorFactory {
-
-	public abstract RadioCommandExecutor createRadioCommandExecutor();
-	
-}
