--- src/main/java/radiounit_executorandreceiver/RadioUnitReceiverFactory.java
@@ -1,12 +1,0 @@
-package radiounit_executorandreceiver;
-
-/**
- * Abstract class for RadioUnitReceiverFactory classes
- * 
- * @author esiumat
- *
- */
-public abstract class RadioUnitReceiverFactory {
-
-	public abstract RadioUnitReceiver createRadioUnitRecevier();
-}
